requirements.txt
streamlit>=1.22.0
sacrebleu>=2.3.1
python-Levenshtein>=0.21.0
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.12.2
sentence-transformers>=2.2.2
torch>=2.1.0
transformers>=4.37.0
